pyproject.toml
[project]
name = "tender-rag-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"  # Compatible with GitHub Actions runners
dependencies = [
    "charset-normalizer>=3.4.4",
    "fastapi>=0.124.2",
    "fasttext>=0.9.3",
    "ftfy>=6.3.1",
    "lingua-language-detector>=1.4.2",
    "pdfplumber>=0.11.8",
    "pymupdf>=1.26.6",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.30.3",
    "sqlalchemy>=2.0.45",
    "greenlet>=3.3.0",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.31.0",
    "supabase>=2.25.1",
    "jinja2>=3.1.6",
    "pymilvus>=2.4.4",
    "requests>=2.32.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sphinx-design>=0.6.1",
    "neo4j>=6.0.3",
    "python-dotenv>=1.2.1",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "linkify-it-py>=2.0.3",
    "myst-parser>=4.0.1",
    "sphinx>=8.2.3",
    "sphinx-autodoc-typehints>=3.5.2",
    "sphinx-copybutton>=0.5.2",
    "furo>=2025.12.19",
    "sphinxcontrib-mermaid>=1.2.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
